realtime updates for invites

+++ lib/actions/invite.ts
@@ -2,33 +2,110 @@
 
 import { revalidatePath } from "next/cache";
 import { db } from "../db";
-import { $Enums, Invite } from "@prisma/client";
+import { auth } from "@clerk/nextjs";
+import { ActionResponse, EventInviteData } from "@/types";
+import { getInviteQuery } from "../query-definitions";
+import { pusherServer } from "../pusher-server";
+
+export async function getEventInviteData(
+  eventId: string
+): Promise<ActionResponse<EventInviteData>> {
+  try {
+    const event = await db.event.findUnique({
+      where: {
+        id: eventId,
+      },
+      include: {
+        invites: {
+          include: {
+            createdBy: {
+              include: {
+                person: true,
+              },
+            },
+          },
+        },
+        memberships: true,
+      },
+    });
+
+    if (!event) {
+      return { error: "Event not found." };
+    }
+
+    const { userId }: { userId: string | null } = auth();
+
+    if (!userId) {
+      return { error: "User not found." };
+    }
+
+    const membership = event.memberships.find(
+      (membership) => membership.personId === userId
+    );
+
+    if (!membership) {
+      return { error: "You are not a member of this event." };
+    }
+
+    if (membership.role === "ATTENDEE") {
+      return { error: "You do not have permission to view this page" };
+    }
+
+    return { success: event };
+  } catch (error) {
+    console.log(error);
+    return { error: "Unable to get invites." };
+  }
+}
 
 export async function createInvite({
   name,
   eventId,
-  createdById,
   maxUses: uses,
   expiresAt,
 }: {
   name?: string;
   eventId: string;
-  createdById: string;
   maxUses: number | null;
   expiresAt: Date | null;
 }) {
   try {
+    const { userId }: { userId: string | null } = auth();
+
+    if (!userId) {
+      return { error: "User not found" };
+    }
+
+    const membership = await db.membership.findFirst({
+      where: {
+        personId: userId,
+        eventId: eventId,
+      },
+    });
+
+    if (!membership || membership.role === "ATTENDEE") {
+      return { error: "You do not have permission to create invites." };
+    }
+
     await db.invite.create({
       data: {
         name: name,
         eventId: eventId,
-        createdById: createdById,
+        createdById: membership.id,
         expiresAt: expiresAt,
         usesRemaining: uses,
         maxUses: uses,
       },
     });
     revalidatePath("/");
+
+    const queryDefinition = getInviteQuery(eventId);
+    pusherServer.trigger(
+      queryDefinition.pusherChannel,
+      queryDefinition.pusherEvent,
+      { message: "Data updated" }
+    );
+
     return { success: "Invite Created" };
   } catch (error) {
     console.log(error);
@@ -38,8 +115,48 @@ export async function createInvite({
 
 export async function deleteInvite(id: string) {
   try {
+    const { userId }: { userId: string | null } = auth();
+
+    if (!userId) {
+      return { error: "User not found" };
+    }
+
+    const invite = await db.invite.findUnique({
+      where: {
+        id: id,
+      },
+      include: {
+        event: {
+          include: {
+            memberships: true,
+          },
+        },
+      },
+    });
+
+    if (!invite) {
+      return { error: "Invite not found." };
+    }
+
+    const membership = invite.event.memberships.find(
+      (membership) => membership.personId === userId
+    );
+
+    if (!membership || membership.role === "ATTENDEE") {
+      return { error: "You do not have permission to delete this invite." };
+    }
+
     await db.invite.delete({ where: { id: id } });
+
     revalidatePath("/");
+
+    const queryDefinition = getInviteQuery(invite.eventId);
+    pusherServer.trigger(
+      queryDefinition.pusherChannel,
+      queryDefinition.pusherEvent,
+      { message: "Data updated" }
+    );
+
     return { success: "Invite Deleted" };
   } catch (error) {
     console.log(error);
@@ -49,8 +166,48 @@ export async function deleteInvite(id: string) {
 
 export async function deleteInvites(ids: string[]) {
   try {
+    const { userId }: { userId: string | null } = auth();
+
+    if (!userId) {
+      return { error: "User not found" };
+    }
+
+    const invites = await db.invite.findMany({
+      where: {
+        id: { in: ids },
+      },
+      include: {
+        event: {
+          include: {
+            memberships: true,
+          },
+        },
+      },
+    });
+
+    if (!invites) {
+      return { error: "Invites not found." };
+    }
+
+    const membership = invites[0].event.memberships.find(
+      (membership) => membership.personId === userId
+    );
+
+    if (!membership || membership.role === "ATTENDEE") {
+      return { error: "You do not have permission to delete these invites." };
+    }
+
     await db.invite.deleteMany({ where: { id: { in: ids } } });
+
     revalidatePath("/");
+
+    const queryDefinition = getInviteQuery(invites[0].eventId);
+    pusherServer.trigger(
+      queryDefinition.pusherChannel,
+      queryDefinition.pusherEvent,
+      { message: "Data updated" }
+    );
+
     return { success: "Invites Deleted" };
   } catch (error) {
     console.log(error);
@@ -121,6 +278,13 @@ export async function acceptInvite({
       where: { id: inviteId },
       data: { usesRemaining: { decrement: 1 } },
     });
+
+    const queryDefinition = getInviteQuery(invite.eventId);
+    pusherServer.trigger(
+      queryDefinition.pusherChannel,
+      queryDefinition.pusherEvent,
+      { message: "Data updated" }
+    );
 
     return { success: "Invite successfully used" };
   } catch (error) {
